+++ src/lib/djerba/plugins/plugin_tester.py
@@ -51,7 +51,8 @@ class PluginTester(TestBase):
                 raise RuntimeError(msg)
         return plugin_name
 
-    def run_basic_test(self, test_source_dir, params, plugin_name=None):
+    def run_basic_test(self, test_source_dir, params,
+                       plugin_name=None, log_level=logging.WARNING):
         """
         Simple plugin test
         """
@@ -61,14 +62,14 @@ class PluginTester(TestBase):
         if not plugin_name:
             plugin_name = self.read_plugin_name(ini_path)
         self.assertTrue(plugin_name)
-        djerba_main = core_main(self.get_tmp_dir(), log_level=logging.WARNING)
+        djerba_main = core_main(self.get_tmp_dir(), log_level=log_level)
         config = djerba_main.configure(ini_path)
         config.set(core_constants.CORE, core_constants.REPORT_ID, 'placeholder')
         data_found = self.redact_json_data(djerba_main.extract(config))
         with open(expected_json_path) as json_file:
             plugin_data_expected = json.loads(json_file.read())
         plugin_data_found = data_found['plugins'][plugin_name]
-        validator = plugin_json_validator(log_level=logging.WARNING)
+        validator = plugin_json_validator(log_level=log_level)
         self.assertTrue(validator.validate_data(plugin_data_found))
         self.assertEqual(plugin_data_found, plugin_data_expected)
         # TODO check other document types, eg. research